requirements.txt
Pillow==9.0.1
python-magic==0.4.25
